+++ src/main/java/cn/nukkit/scoreboard/scorer/FakeScorer.java
@@ -41,6 +41,6 @@ public class FakeScorer implements IScorer {
 
     @Override
     public SetScorePacket.ScoreInfo toNetworkInfo(IScoreboard scoreboard, IScoreboardLine line) {
-        return new SetScorePacket.ScoreInfo(line.getLineId(), scoreboard.getObjectiveName(), line.getScore(), getFakeName());
+        return new SetScorePacket.ScoreInfo(line.getLineId(), scoreboard.getObjectiveName(), (int) line.getScore(), getFakeName());
     }
 }
